feat: complete back-end / front-end interaction

+++ app/Http/Controllers/ExportController.php
@@ -5,6 +5,7 @@ namespace App\Http\Controllers;
 use App\Models\Export;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Storage;
+use Inertia\Inertia;
 
 class ExportController extends Controller
 {
@@ -13,18 +14,35 @@ class ExportController extends Controller
      */
     public function index()
     {
-        $exports = Export::paginate(15);
+        $exports = Export::paginate(5);
+
+        return Inertia::render('Reports', [
+            'exports' => $exports
+        ]);
+    }
+
+    /**
+     * Display the specified resource.
+     */
+    public function show($export) {
+        $export = Export::find($export);
+
+        return Storage::download($export->file_name);
     }
 
     /**
      * Remove the specified resource from storage.
      */
-    public function destroy(Export $export)
+    public function destroy($export)
     {
-        Storage::delete($export->file_name);
+        $export = Export::find($export);
 
-        $export->delete();
+        if($export) {
+            Storage::delete($export->file_name);
+            $export->delete();
+        }
 
-        return '[INFO] All the informations are destroyed!';
+        return redirect()->back()
+            ->with('success', 'Arquivo removido com sucesso!');
     }
 }
